add home page erb file

+++ db/migrate/20220605014638_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[5.2]
   def change
     create_table :products do |t|
       t.column(:name, :string)
-      t.column(:cost, :integer)
+      t.column(:cost, :decimal)
       t.column(:country_of_origin, :string)
 
       t.timestamps()
